update code to gabe's recent refactorings
